Add validation that MW start/end cannot be in past
Unless the MaintenanceWindow is in a state where these are allowed to be
in the past.

Currently the whole maintenance period can be passed for a
MaintenanceWindow which is in a 'finished' state (though this will
remain the case once `expired` is no longer considered a 'finished'
state, in the near future), and the `requested_start` only can also be
passed for `started` MaintenanceWindows.

These restrictions on these validations are necessary to avoid existing
MaintenanceWindows saved in the database eventually becoming invalid,
once their maintenance periods have passed, which would likely have
caused us issues in the future.