Support transitioning MWs out of expired state
We have now been requested to support transitioning MWs out of being
`expired` in a similar way to if they were still `requested`.

This is intended to make the process of requesting maintenance more
straightforward to users, as all maintenance should explicitly be
progressed to be either `confirmed` or `cancelled`/`rejected` via the
actions of some user, rather than becoming `expired` after which it
could no longer be able to be transitioned and maintenance for the same
model would then need to be re-requested (which would likely cause
confusion).