Rename MW `skip_comments` flag to be more general
This will now tweak various behaviour we do not want to occur when
migrating all existing MaintenanceWindows to be compatible with the new
way maintenance will work. Once our production data has been migrated we
will then be able to remove this flag as we should not need the special
handling it enables again.