Skip MW future validation when migrating data
When migrating our current production data to the new MaintenanceWindow
system we simulate transitioning MaintenanceWindows through the
different states they would have passed through if we were always using
the current system. This works fine, apart from that most of this
maintenance should have already occurred in the past and so many of
these transitions are now invalid (due to recently added validations).

This commit allows us to skip just these validations only when we are
migrating this legacy data; after we have done this in production we
will then be able to remove this special handling.