Remove the helper method for rendering tabs
Now that tab-bar use the `scope` to retrieve which tabs to use, there is
no longer a need for the helper method.

The blank card partial can also be removed, the tabs bar can be rendered
for models using the base decorator. This did require the decorator
`tabs` method to be switched to public.