Have maintenance form POST back to same URL
This has the advantage that if the form submission fails the user will
remain on the same URL and will not then be surprised if they then
refresh the page and have it entirely change. In particular this will be
more useful shortly once we gracefully handle failed form submissions
and re-render the new maintenance form with all errors shown.