Create a show path for the asset_record
The asset_record shall be shown in its own tab and thus a `show` action
needed to be added.

A route to this new `show` action has been added for
`component_groups`.

Work is still required to properly integrate the `show` action with
tabs. Currently it is double rendering the header.