Remove the show page for ComponentGroups
Now that `asset_record` and `components` will have there own tabs, the
show page for `ComponentGroups` is now redundant.

However as there are existing links internally (and possible externally)
to this page, it can't be removed completely.

Instead, the `show` path still exists BUT it will render the
asset_record instead. This way there will not be any issues with the
publicly available path disappearing.