DRY up the creation of the overview tab
Both the Cluster and Component pages use the `Overview` tab as there
landing page. Thus the creation of the tab has been moved into
`TabBuilder`.

The `TabBuilder` needed to be adjusted so it can render paths without a
method. It might be a good idea to extract the `scope_path` method out
into a general `helper`.

It would also be good if `scope_asset_record_path` could be
automatically converted to `component_asset_record_path` using gsub.
This way the helper would be a lot more versatile.