Refactor DateTimeSelect to take model + field name
This is a bit simpler, as we now only need to pass the datetime field
name once rather than twice in slightly different forms, and will also
be more flexible as we have access to the actual model object within a
DateTimeSelect instance. This will be useful soon for determining if the
model has any validation errors and altering how the selects are
displayed based on this.