Split out maintenance tests to test one thing
Rather than having just a couple of large tests with a jumble of
interacting with the page and assertions.

Some tests still have many assertions but that is because these are to
all test the same aspect of interacting with the form, and I think it
would be impractical (and much slower to run) to split these out
further.