Have MW confirm button link to new confirm page
Rather than having maintenance just be confirmed by users with no
ability to alter it, we want to have them confirm this by visiting a new
confirmation page, similar to the current request maintenance page,
where they can alter the dates for the requested maintenance.

This commit handles creating and appropriately linking to this route and
page; the form itself is still to implement.