Pass MW as argument to form partial
Rather than relying on a controller instance variable having been set.
Passing this as an argument is better as it makes this dependency of the
partial explicit, and also means things should fail obviously if this
argument is ever not given, rather than just attempting to use a `nil`
instance variable.