Make MW#create use transaction more effectively
Previously we did not use any 'bang' methods in the transaction, which
would raise an exception if the updates failed, so the transaction would
not rollback in the standard situation where we would want it to (when
either the MaintenanceWindow creation or requesting failed).

This does not seem a particularly big deal at the moment as I can't see
a situation where one of the database updates would fail but the other
wouldn't, but it still seems a good idea to ensure these continue
happening together or not at all. This commit fixes this so this should
now be the case.