Remove assertion that requested_start has no error
When testing our standard error handling when the maintenance form is
submitted, we should not assert anything about the validity of
`requested_start` as we are not explicitly changing it in the test so it
may or may not be valid depending on the factory used. This will shortly
be relevant when testing expired maintenance, as this test otherwise
works fine in this situation.

If we want to explicitly test error handling when changing
`requested_start` then we can add a new test to do so later.