Make confirming expired maintenance work
Should work very similarly to confirming non-expired requested
maintenance. Specific changes needed to handle this correctly:

- altering how we determine what text to display for form submit button,
so it displays correct text in this case;

- update the MaintenanceWindows requested start/end and state in a
single update, rather than in multiple updates wrapped in a transaction,
either of which could fail with inconsistent errors in this case, which
ensures we display the same possible errors in this case;

- handle and display errors given when attempt an invalid state
transition, as this is the error which will now occur if the update
fails.