Remove the titles spread across the controllers
Previously the titles where being independently set across all the
controllers. This has lead to the inconsistency in what the title should
be. This commit removes the titles so they can be set in a single
location.

The only title that hasn't been removed is for the `admin` root page.
It's title is a bit different and might need to be addressed differently
(maybe?).