Fix test so doesn't fail on a Monday o_O
We now calculate the `expected_end` for a MaintenanceWindow as
`duration` business days after the `requested_start`. This had the
unfortunate side effect of making this test that a `started`
MaintenanceWindow is correctly transitioned to `ended` fail as
`expected_end` is not actually passed when `2.days.ago` falls over a
weekend. A quick fix for this is to just make `requested_start` longer
ago so `expected_end` should always be in the past.