Assign the title in the ApplicationController
This ensures a consistent title is used across all the pages. Currently
the show `case` page is not within a scope and thus `scope.name` will
error. As such a guard is required to protect against this.

Also the `define_navigation_variables` has been renamed `assign_scope`
as this is its actual function