Remove the subtitles and de-duplicate the scope assignment
The `subtitles` are going to be assigned in the view to make them
consistent across the tabs. This commits removes the subtitle assignment
from the contollers.

Also a lot of code has been removed from the `cluster` and `services`
controllers. Some of it was to do with the now obsolete `subtitles`.
However the controllers where also duplicating the `scope` assignment
from the `ApplicationController`. This has now been removed as well.