Switch the subtitle input from a variable to a yeild
Instead of assigning the subtitle using a instance variable,
`content_for` will be used in the views. To facilitate this, a yield
needs to be added to the application layout.

Because not all pages will have a subtitle (possible), a check needs to
be performed to check for empty strings.