Standardize the subtitle texts
The subtitles on the index pages do not need to say "view" in the text.
It is clear to the user that they are viewing a list. Instead, the
subtitle is what they are viewing (e.g. `components` or `expansions`).

Also on the show pages, they should only say `Overview` instead of
`Component Overview`. This could go either way, however no other
subtitles have the scopes class in the name.