Determine comment to use based on event not state
Rather than using the state of the MaintenanceWindow to determine what
comment should be left regarding the most recent state transition on the
associated RT ticket, use the event of the transition itself.

This has the benefit that we can change which comment is left based on
how the new state was reached, which will be useful shortly when a new
way to transition a maintenance window to `confirmed` is added which
bypasses the usual confirmation process.