Append the Component type in the subtitle
This shows which list of components are currently being viewed. The type
has to be set in the controller as it is part of the request. However to
ensure that there is always a subtitle, the type is append from within
the view.

This means if the view is rendered from somewhere else, than the basic
`Components` subtitle will still be rendered.

https://github.com/alces-software/alces-flight-center/pull/138#discussion_r177180738