Update the scope_path method to work with sites pages
The site pages are a bit tricky as they have a different route depending
if the current_user is an admin or not. If they are a admin, then it is
the standard `/scope/:id/...` path.

However when a contact is signed in, the do not have the `/site/:id`
paths as they only belong to a single site. Instead the root path takes
the place. This makes forming the path helpers a bit trickier as they
need to make a distinction between the two cases.

Also added a spec for ensuring the correct paths are used for each type
of user.

This did require updating the spec so that it matches the new paths