Add an `archive` route along side the index cases
When indexing the cases, there is two modes: regular and archive.
Previously the archive was done by sending a `archive=true` parameter
with the request.

However as this is a boolean, the toggle can be included in the route
itself. This commit adds a `/cases/archives` route for every `/cases`
(index cases).

This new route will eventually replace the request parameter. A proc has
been used so it can be easily reused in multiple places. NOTE: The
`/cases/archives` route must be defined first so it is not confused with
the `/cases/:id` route which matches the show action.

Note that the `archive` action will archive an individual case, where
the `archives` action will retrieve all the cases.