Add scope path helper for helper finding paths
Now that there is a scope that changes, it makes finding the paths to
different pages harder. A new helper has been added to make this process
easier.

The ScopePathHelper, uses method missing to convert paths which contain
the phrase `_scope_` to the corresponding path. Some additional logic is
required to handle when a path helper starts with `scope_`.

The scope is always assumed to be the first argument input. For example
`scope_path(some_cluster)` is converted to:
`cluster_path(some_cluster)`.