Move the Site specific code into the SiteDecorator
When a contact is signed in, the site path is not included in the URL.
Instead it needs to return a underscore so the path is correctly
formatted without it.

Instead of preforming this check every time, inheritance is used so it
is only preformed when the scope is a `Site`.

The spec has been updated so it doesn't set the `current_user` for every
test. Instead it only needs to do this when the `scope` is a Site. This
makes it clearer what is required for each spec.