Do not pass the site model when it is not required
As contacts use the root paths instead of the site paths, there are
situations the site model isn't used. If it is however passed anyway, it
will break the routes.

Instead, only pass in the raw inputs if the scope is a `Site` and the
current user is a contact.