Toggle all transition validations based on event
We have several validations for MaintenanceWindow transitions which
should only be run depending on the particular type of transition.
Rather than having some of these validations be enabled/disabled
depending on the state a transition is transitioning to, as was done
previously, have this be done based on the transition `event`.

This is better as we may have multiple `event`s which can transition to
the same state but we may want to validate properties differently
depending on the particular event used.

Also, while doing this, I have switched so that the white-listed events
for which we check for whether a user is present/absent are the
automatic events, rather than the user-initiated ones. This was done as
this list is shorter and less likely to change, while it is likely to be
more frequent that we add new events which can be triggered by users.