Prefix automatic MW events with `auto_`
This has the advantage of easily distinguishing all of these events from
the other, user-initiated events.

This was prompted by needing to provide events which can be triggered by
users for the same kind of transitions as these, in particular we want
to provide an `end` event which can be triggered by admins to perform
the same transition as the `auto_end` event (previously just `end`).
Having this distinction should make this straightforward to do, while
allowing us to have both distinct messages in both cases and sensible event
names.