Change the ComponentGroups landing page to index Components
Currently as there is no Overview for ComponentGroups, a different page
is required for the landing page. Previously it was the `AssetRecord`
however a list of components is the more useful page.

For the Components list to be rendered correctly, the Components
controller must be used (instead of the ComponentGroupsController).
To do this, the routing has been updated to send the request to the
ComponentsController#index action.