Simplify regex/sub used for scope_path methods
Previously only `_scope_` was being matched. This meant methods that
started with `scope_` had to be pre-processed to contain `_scope_`
before the `sub` could occur. This also meant post-processing was
required to remove the leading `_`.

Instead a capture group can be used to match all text up to the first
underscore in `_scope_` OR match the beginning of the string. This
capture group is then included in the replacement using the `\1`. This
removes the need for pre and post processing.

The handling of the trailing underscore has been altered as well.
Previously the `scope_name_for_path` needed to handle it. However this
gets confusing for the root paths. Instead the underscore gets append
automatically if a name is provided (as opposed to empty string which
denotes the root path).

Addresses: https://github.com/alces-software/alces-flight-center/pull/141#discussion_r178788803