Add Tiers and columns to track Tier data to Cases
Every Issue will have 1+ Tiers.

Every Tier is at a particular `level` and contains some configuration
JSON for its possible `fields`.

Every Case will be created for a particular Tier and so contains a
particular `tier_level` and the completed `fields` JSON (which will
contain the same data as the `fields` for the Tier that it was created
for, but should also have `value` properties set for each entry in
`fields`).

Note that we do not want to link Cases to Tiers directly as the Case
represents a particular support Case that was created for an available
Tier at a particular moment in time, however any Tier could later be
removed or changed to have a different `level` or different `field`s, so
it makes sense to just take a snapshot of the relevant Tier data at the
moment the Case is created.