Initial migration to update Issues/Cases for Tiers
Sketch out the initial data migration to update all current Issues and
Cases for the new relations to Tiers or Tier-related fields,
respectively, that they will now need.

The exact details of the specific Tiers or the format of the
Tier-related fields that we need may later change as we determine our
requirements in more detail, or if we alter the data format needed for
Tier fields; if this happens this migration can then be updated or an
additional migration made as needed (depending on whether we have
already deployed it).