Thread through Field type to generate field labels
Now that we have a type which represents all the possible fields in the
Case form, we can thread this through each function used to generate the
corresponding input and `toString` it to get a label, rather than
passing through the label directly.

This will also be useful shortly when using this Field value to validate
fields, and means we won't need to pass in, for example, both
`Validation.Cluster` and `"Cluster"` to the same function.