Decode junk Tiers as additional layer below Issues
Tiers will appear as another SelectList below Issues, and
correspondingly be selectable via another `select` shown under the Issue
`select`. This commit is the first step towards this, by decoding a
(currently hard-coded) SelectList of Tiers for Issues; this SelectList
is also ordered by the level of each Tier (so more straightforward
solutions to problems, like guides, will be selected and shown first).