Make Case form UI slightly clearer with Tier fields
Add a separator between the (mostly, as Category select is only
sometimes shown) static upper select fields and the dynamic lower
Tier-specific fields, to make slightly clearer that the lower fields are
distinct and will change based on the selection of the upper fields.

Also move component select and subject fields into the lower section, as
these are only sometimes shown and dynamically update based on the
selected issue, respectively, and so appear to fit better here.