Encode Case details from Tier fields
Rather than encoding the details for the Case to be created from the
`details` field of the selected Issue, create this using the entered
values from all fields of the selected Tier.

This is an improvement over what was done previously and means we can
now remove the Issue `details` field entirely. Longer term the Case will
be created via the Tier fields using JSON rather than just a `details`
string, so that we can retain knowledge of the structure of the Tier
fields in the created Case.