Remove Issue details field from Case form
A Case is now created by filling in a number of fields specified by the
selected Tier, rather than by filling in a single 'Details' `textarea`
where anything can be entered, which was pre-filled using the selected
Issue's details template.

Therefore we can now remove all references to the Issue details field
(apart from when encoding the data to create the Case with, as for now a
single `details` field is still formed from the current Tier fields to
conform to the format the CasesController expects).