(Mostly) remove Issue details_template field
This is no longer needed/used anywhere, so can be made nullable and most
references to it removed. We still need to keep the field itself around
for now so that we can read and migrate the data within it; once this
has been done on the next deploy then the field can be removed entirely.