Encode and save Tier data on Case form submit
We now include the Tier level and fields when submitting the Case form,
and save these in the created Case.

Only Tier fields which may have a value associated with them are
included in the submitted JSON, and of those fields which are submitted
the value is also included in the field object (which is the only
additional data included over when the fields are initially decoded in
the Case form app).