Fix issue with creating Case for advice-only Issue
In the particular situation where an attempt is made to create a Case
with:

- an `advice-only` Issue which requires a Component or Service;

- a `managed` Cluster;

- an `advice` Component or Service (as appropriate to match the Issue);

the Case would be invalid with the error that an advice-only Issue
cannot be created for a managed Cluster.

However, this is surprising and not what we want to happen, as just the
Component/Service's `support_type` should be considered in this
situation to determine if the Issue can be associated. This changed
behaviour is also consistent with what we already did when creating a
Case with a `managed` Issue which requires a part and an `advice`
Cluster.