Add needed fields to support_type Issue factories
These factories need to create Issues with the fields expected when we
use these special Issues (i.e. appropriate values for
`requires_component`, `requires_service`, and `support_type`), so that
when we use these in tests we are automatically testing with values
similar/the same as will be used in production (at least unless the
production values are independently changed).