Switch to feature tests for support_type buttons
The previous tests of the decorator methods were simultaneously brittle,
as they checked for various specific details of the markup which
shouldn't matter to tests, and also provided inadequate coverage, as
these buttons had now stopped successfully creating Cases following
recent changes to the CasesController, but the tests still passed.

This commit replaces these with feature tests which actually test that
the buttons work correctly, without testing unimportant aspects of the
markup, and also updates the buttons to actually work correctly
according to these tests and to include the new data expected by
CasesController.