Move Component select field into top form section
An argument could be made for putting this field in either of the form
sections we now have, since it is both dynamically shown based on
whether the selected Issue requires a Component (as with the rest of the
lower section) but could also be considered part of drilling down to
determine the reason for creating a Case (as with the upper section).

Previously it was in the lower section for the reason described above;
however, since we now want to disable the lower section based on the
`support_type` of both the currently selected Service and Component (if
applicable), it makes sense to move it to the upper section so the
entire lower section can now be disabled when needed.