Add support for specifying String error messages
Previously we only supported specifying error messages as `Empty`, which
is used to indicate that a Field has an error where the cause is clear
and can be easily addressed (e.g. for empty fields). This commit expands
this to also allow specifying an error message with accompanying String
for the less obvious cases.