Disable lower form section when Tier unavailable
When the currently selected Tier is not able to be selected (based on
which Tier is selected and whether this is compatible with the selected
Component or Service) we want to:

- disable the fields in the lower section of the form;

- make this whole section appear unavailable, by fading it out and
adding title text indicating why it is unavailable (with the same error
message as shown for the Tier itself).

The reason for this section being unavailable is also already indicated
by the Tier error message.