Always show all Tiers
To make clear to users the possible available support types, rather than
hiding those which are unavailable for the currently selected Issue.
Soon the unavailable Tiers will be shown disabled and with appropriate
text to indicate this.

This is done using a wrapper around Tiers just in the view, so that the
real available Tiers for each selected Issue are still the only ones
stored in the State.