Use association where possible in factories
Previously we did things like `user { create(:admin) }` in various
places in our FactoryBot factories. This mostly appeared to work fine,
as in our tests we mostly `create` things anyway (maybe we should do so
less and use `build`/`build_stubbed` instead, but currently this is what
we do).

However, this has the flaw that the associated object is always created,
and this can cause problems if things are left in the database, for
example unique constraints can then fail. In particular this started
causing an issue as `FactoryBot.build` is now used in
`app/mailer_previews/case_mailer_preview.rb`, and the second time this
line was hit it was causing the email uniqueness validation on Users to
fail due to two users with the same email having been created (even
though `build` was used so the CaseComment itself is not saved).

The solution is to use `association` where possible, so that associated
objects are created using the same strategy
(`build`/`create`/`build_stubbed` etc.) as the original factory.

Note that we still `create` things directly in a few circumstances when
creating factories, as it was not immediately obvious how to remove
these, but none of these look like they will cause us issues soon.