Prevent comment email preview erroring
Use `FactoryBot.build_stubbed` to create a CaseComment which looks like
it has been persisted, with fields like `updated_at` set, rather than
`FactoryBot.build`, which does not. This prevents viewing the
CaseComment email preview erroring with a fresh database (without
existing comments, so this branch is hit), as `updated_at` is accessed
in this email template.