Don't log initial state transition for `Case`.
There's no need to do so, since the data is already captured as the
`user` of the case and its `created_at` date. We don't want to create a
separate event card for it.

Also, the state transition can be created before the `user` gets
assigned, in which case we might have no idea of which user caused the
transition.