Wrap `aws` CLI so uses correct profile
For some reason, which I don't have the inclination to debug in detail
right now, this script was failing due to invalid credentials on my new
machine when run from a Rake task (as in
`alces:data:import_production`), although it both works when run outside
a Rake task and `aws` appears to be using the correct credentials when
checked when running within a Rake task.

Explicitly specifying the profile I want fixes this for me, and
hopefully doesn't break it for anyone else.