Remove validation of Case-part based on Issue
We previously had a lot of in depth validation about whether the
Case-part (Component, Service, or Cluster, from highest to lowest
priority) relationship is valid based on the `support_type` of both the
part and the Issue.

With the addition of Tiers this is no longer needed, and we will instead
soon replace this with a simpler scheme where a Case can only be created
for a level 3+ Tier if it is associated with an `advice` part. Other
than this restriction a Case will be able to be created for any part and
Issue, as it will be the chosen Tier level which determines when support
can be created rather than a property of the Issue itself, and
eventually Issue `support_type`s will be able to be removed entirely.