Extract Case#text_summary tests from CaseMailer
In preparation for improving how we generate the text for Case emails,
to include the new Tier fields and to make these slightly nicer
formatted.

The full tests, including for edge cases with optional fields, have been
extracted to the Case specs, while we still test the overall email body
format is what we expect in the CaseMailer specs to ensure this is being
correctly created.