Move formatting Case email into mail templates
Rather than generating a formatted string in `Case#text_summary` and
then munging this as needed in the two Case email templates to get
things in a suitable format for display in emails, we now generate the
properties we would like to appear in both email formats and then handle
appropriately formatting this in the emails themselves.

This allows us to have slightly more control over the formatting of
these properties, as they do not have to be shown as mostly plain-text
in both cases.