Hide credit charge-related fields from Case pages
These are no longer meaningful, and are now misleading, as we

a. are not currently charging credits for support within Flight Center;

b. are now considering Cases 'chargeable' based on the Tier they were
created for, rather than the `chargeable` property of the Issue they
were created for (which will likely be removed at some point).