Improve multi-line Tier field formatting in emails
For multi-line Case fields we want to preserve the original formatting
in the email sent on Case creation, in a manner appropriate to the
format, while still showing non-multi-line fields in a compact way.

Also needed to tweak a couple of places in tests where we generate Case
fields, to conform to the expected format.