Prevent unnecessary breaks in MW event cards
We recently started `simple_format`ing the text shown in event cards.
This is mostly what we want, to preserve user-formatting in comments,
but not in this case as this leads to unnecessary line breaks in
maintenance transition cards wherever there happen to be line breaks in
the templates used to generate this text.