Remove mailto link from Case table (fixes #198)
Now that we have (mostly, soon to be entirely) removed our integration
with RT, and have added support for commenting and email sending in
Flight Center, there is no need to provide an easy way for users to
contact us outside of Flight Center, and in fact we want to avoid doing
this. This commit removes the link to do this, as well as the template
used for this link and `Case#mailto_url` method which then became
unused.