Have display_id distinguish between Case types
Show IDs differently wherever we do this depending on whether the Case
is an RT ticket or not. Eventually https://trello.com/c/dzY3fb5C will be
implemented and every Case will have an identifier like `RT12345` or
`BAR123` (where `BAR` is unique to the particular Cluster), which will
make it easy to determine the type of Case at a glance.