Show historical comments and Logs with newest first (fixes #200)
In the case of CaseComments the comment box is also shown at the top of
Case history, since it fits well here. For Logs this has not been done
as:

a. it would be slightly trickier and probably look worse, as the add log
form is currently in a separate card and has not been integrated into
the tabbed page structure we now use;

b. this form is only shown to admins so we don't care as much at the
moment;

c. we will likely refactor this UI properly at some point to fit better
in the tabbed interface, making a minor tweak like this unnecessary.