Use display_id in ticket_link (fixes #193)
We previously hadn't adapted this method to account for the 2 types of
Cases we now have, those with and those without an associated RT ticket.

This commit does this in a way which makes clear what kind of Case we
have, by using the `display_id` in this method; the method has also been
renamed since it is no longer ticket-specific.