Add new fields to Cluster: `shortcode` and `case_index`.
In the data migration we make a token effort to deduplicate candidate
shortcodes for clusters; in practice this code probably won't be
exercised that often, but might help smooth the migration on dev
databases that might have a bunch of weird data in.

We can't yet make the `shortcode` field mandatory in the database,
because it requires the data migration to be run first in order to not
violate that constraint. We do, however, enforce its presence at the
model level.