Add first-class `display_id` to `Case`.
Includes a data migration to populate `display_id` for all existing
`Case`s. This will take one of two forms:

- `RT12345` for a `Case` with an `rt_ticket_id` of `12345`;
- `FOO12345` for the 12345th `Case` in the Foo `Cluster`.

Some minutiae:
- We use a database lock when generating the per-cluster numbering in
`Cluster#next_case_index` to ensure that concurrent accesses can't end
up with the same value
- We only validate the presence of a `display_id` after the record has
been persisted. This means that we check a Case is (otherwise) fully
valid and saved before we increment the cluster's counter, and thereby
ensure that cases are generated with sequential display IDs.