Remove ticket status update for admins (fixes #197)
Doing this when an admin visits a `/cases` or `/cases/resolved` page,
especially given our increasing number of non-resolved Cases, was
causing a massive slow down in this request, due to us needing to make a
large number of sequential requests to our tiny RT server.

We could try to improve how we do this, but since we plan to remove all
integration with RT entirely soon, and don't need to keep this part for
the transition, let's just remove it now.