Test for the associated component of a log
Adds to the existing component tests when visiting the cluster log with
and without a component associated to a log. These tests should ensure
that the page contains 'None' for a log without a component. But a link
with the name of the component, pointing to it's respective component
page for a log that does have a component.