Set a fixed width for fields within cases
Without a fixed width the table can look messy, especially with little
information added to each field. The user would need to look over a lot
of empty space before seeing anything relevant which isn't a good use of
screen economy.

This commit sets an explicit width of 25% on the fields so that the
values of each field have the majority of the space reserved for them.