Implement simple state machine for ChangeMotdRequest
At this point this is possibly slightly overcomplicated for what we need
it for, since ChangeMotdRequest can just be `unapplied` or `applied`,
and either way it can have `apply` run on it, and we also want to track
the transitions it undergoes.

However I thought it was still worth doing this in this case, firstly as
we are considering later tracking various additional states so this will
be easy to extend, rather than having to migrate existing data later
(see https://github.com/alces-software/alces-flight-center/pull/226).
Secondly since, even though this is simple, it gives us various useful
behaviour for free/cheaply, such as allowing us to query states, have
the transitions with the user who initiated them recorded etc., which we
might as well take advantage of.