Display and update MOTD tool fields
The fields for the MOTD tool, which currently just consist of a single
text area for editing the Cluster MOTD, can be handled in a similar way
to the fields for any other Tier, except that these are defined within
the Case form app rather than the database.

This is done for two purposes: firstly so we have more control over
these fields within the Case form app (i.e. so the initial value for the
MOTD field can be set to the current MOTD for the Cluster), and secondly
to provide an initial example for how things will work once we add more
complex Tier 1 tools (the User Management tool etc.).