Change the styling for current user highlighting
Cases assigned to the current user are now highlighted with a pale blue
background. The old gray background could be seen as designating the
cases as not as important to some users. This new colour still fits the
colour palette and indicates these cases are different in some form.

On top of this the name of the assignee is emboldened for these cases.