Use implicit decoration when forming Service JSON
Rely on various uses of `decorates_association` so associations are
automatically decorated.

Some uses of `decorate` are still required, for models which are not
associated to the Service we are forming JSON for, but possibly I like
that this is explicit just for these models, and is otherwise automatic
for those which are directly associated.