Treat `clusters` like an array, because it probably is one.
In three of the four ways that `clusters` can be set, it's set to an
array (a singleton, as it happens). Only in the fourth case is it set to
an `ActiveRecord::Associations::CollectionProxy` with a `decorate`
method, and that was the only case in which the new case form was
working (when accessed from a site dashboard).

This commit fixes the form to work for the other code paths by assuming
`clusters` is an array, and calling `decorate` on each of its members
directly.

Fixes #254.