Force table headers to exist on a single line
A couple of table headers present contain more than one word separated
by a space which causes them to span multiple lines. This has the effect
of pushing the table content down slightly and impacting the overall
look of the table. This commit applies the shrink class to the table row
that contains these headers, forcing them to not wrap onto multiple
lines.