Implement `Case#tool_fields=`
Can now just call `tool_fields=` on a Case with a hash of data, which
ultimately should/will soon come from a particular tool in the Case
form, and have this be appropriately handled by creating the appropriate
associated models or gracefully erroring. This commit implements this,
with (for now) just handling of the `motd` tool and the obvious error
cases; later we can slot more tools into this as they are added.